[FIX] account_accountant: bank reco: no link to unreconciled item
The bank reconciliation widget shows the link to the corresponding
bank journal entry.
Currently this also happens when the entry is not reconciled.
The unreconciled entry is only a placeholder that will be filled with
the "real" information upon reconciliation.
Thus the link to the bank journal entry should not be shown before it
is reconciled (it is not ready yet).

This commit removes the links for unreconciled entries.

task-3425233

closes odoo/enterprise#45355

X-original-commit: ff8e8070627df0d91c626c21211cabcd524afe57
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>