[FIX] account: fix reconciliation of non-receivable/payable account in foreign currency
Reconciling two lines of an account that is not receivable nor payable with one of
the lines being in foreign currency throws an error.
In that case we should reconcile in company currency and not in foreign currency.
Note: we had to rework wizard's computes to avoid cyclic dependencies.

task-3420586

closes odoo/enterprise#44567

X-original-commit: d6450208f041ee26d5dea703b290a57aadc67cfc
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>