[FIX] *: fix grammar error in field help
This commit fixes a really old grammar error in the help message of
message_needaction_counter field.

Before this commit: “Number of messages which requires an action”
After: “Number of messages requiring action”

The subject of “require” is “messages”, which is third-person plural, so
it can't take the -s suffix.

closes odoo/enterprise#44299

Related: odoo/odoo#128929
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>