[IMP] sign: convert sign to OWL
This commit converts the common and backend parts of sign into OWL and removes
jQuery usage in the sign module (note that jQuery is still being used under the
hood as jSignature is based on it, therefore we still have to add it in the
manifest file). Also, some parts of the code that interact with the PDF.js
iframe are not using OWL, but only vanilla JS. This is due to the way we
interact with the PDF.js library. Essentially, we add the sign items to each
page div rendered by the library. If we were to use OWL, we'd have to mount
multiple OWL apps in the iframe and that would be cumbersome to manage. In
order to have the benefits of the webclient and OWL, the solution was to
pass services and the env object to the classes that implement the iframe
logic. In order to keep the iframe logic closer to the experience of writing
OWL code, it was chosen to have all iframe logic be implemented with JS
classes. This also allows us to monkey patch our classes using the patch
function.

task-3286974

Part-of: odoo/enterprise#43103