[FIX] project: fix external id not found issue in project
When the user delete a project in the `Project` module that is present in the
data file of `Field Service - Sale`. While installing or upgrading the
`Field Service - Sale` module, the user gets an error as the project
referenced from the 'Project' module has been deleted.

Steps to reproduce :
1. Install the `Field Service` module.
2. Open the project app.
3. Delete all the tasks of the `Field Service` project and then delete the
   `Field Service` project.
4. Install or upgrade modules that are using that project like
   `Field Service - Sale`, `Helpdesk FSM`, etc.

Traceback will be generated.

This commit will fix the above issue by updating the record that might be
deleted by the user.

sentry - 4087533601

closes odoo/enterprise#45219

X-original-commit: 844868f997b7da2486bee80f0df570b640583b9e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>