[FIX] social_twitter: migrate to the new API
Purpose
=======
Migrate social twitter to the new major version of API, because the
current one is not supported anymore.

Changes
=======
We can not search user by name anymore, we need to type the entire
username directly (so before we could search by typing "Elon Mu",
but now we have to enter the entire screen name, so "elonmusk").

To know if we liked a tweet or not, we need to request a different
endpoint `/2/users/:id/liked_tweets`, but we can not search on this
endpoint, so we will need to iterate over ALL the tweets the users liked
(each time we want to update the like status). So we do not synchronize
the like status with Twitter (the like status is just stored in the
database). See https://developer.twitter.com/en/docs/twitter-api/tweets/likes/introduction

The upload of the attachment uses the v1 API, and it is one of the only
v1 endpoint supported (but we might need to change it at some point,
when a v2 equivalent is available).

With the new API we can only fetch tweets created in the last 7 days.
See https://developer.twitter.com/en/docs/twitter-api/tweets/search/introduction

Task-3431873

Part-of: odoo/enterprise#46656