[IMP] account_asset: Allow to modify the depreciation duration of an asset.
Problem
---------

In V15, we can modify the duration of an asset easily. In V16 we need to
disposal it or to stop it and relaunch it to be able to do that.

Objective
---------

Add the possibility to change the duration of the asset in V16, like
it's available in V15.

Solution
---------

Update the wizard view to display the duration in the modify action.

task-3366584

closes odoo/enterprise#45703

X-original-commit: 098490c4efe78066ee75696263ace12bdb3973d8
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>