[FIX] website_sale_subscription: correct key addition to `t-cache`
Issue:
------
When we go to a recurring product on ecommerce and
want to go to another recurring product,
we stay on the same product.

Cause:
------
The `t-cache` key is overwritten and not added.

Solution:
---------
Syntax correction for key addition on `t-cache` attribute.
`<attribute name="t-cache" add="expr_py" separator=","/>`

opw-3420753

closes odoo/enterprise#44475

X-original-commit: 2caa5eeec0dd751bf145edb89201ec27ba7b2586
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>