[FIX] website_sale_stock_renting: correct daterange picker filter
Issue:
When filtering on rental product with the datepicker for a daterange
in a future period, where the rental product *will* be available,
but it is not *today* because *all* our quantity is being rented out,
the filtering on the `/shop` doesn't return the product in the
search result, even if it should be available for reservation, after
the return of the rental product.

Steps to reproduce:
- Install eCommerce, Rental, Inventory
- Duplicate a rental product, disable stock tracking and disable the
  possibility to sell it out of stock, publish the product on the
  website
- Change the quantity on hand for that product to 1
- Create a new rental order and rent the product for a daterange X->Y
- Pickup the order (quantity on hand is 0)
- On the /shop page, edit the website to add the daterange picker
- Search with the daterange picker a daterange Z->W, where Z > Y
- Observe that our rental product doesn't show up in the search
  result, even if it could be rented out at Z->W

Cause:
The `_filter_on_available_rental_products` filters out variants that
have `qty_available = 0`, without checking that maybe some of them
could be in the process of being rented out.

Fix:
Extend the filtering to take into account that variants could be in
the process of being rented out.

Affected versions:
16.0 up to master

opw-3414160

closes odoo/enterprise#44382

Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>