[FIX] website_sale_stock_renting: fix pricelist issue in tests
Before this commit, the tests were using the website of the main company
to test flows in a new custom company created in the test setup.

This triggered multi-company errors (deterministically or not, depending
on setup)

This commit makes sure the whole setup is restricted to a specific company,
and that the website used belong to that company.

Fixes build error 23056

closes odoo/enterprise#44388

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Co-authored-by: Victor Feyens <vfe@odoo.com>