[IMP] product_unspsc, l10n_dk: Add unspsc code to the danish localization.
Problem
---------
Danish company which works with administration need to use the UNSPSC
code for items.

Objective
---------
Add UNSPSC code usage to the danish localization.

Solution
---------
Add the UNSPSC logic to the DK localization by adding the product_unspsc
dependency in the localization EDI manifest (created for such porpose).

Regarding the translations, this was done in several steps:
    1. Change the UNSPSC field name to translatable
    2. Translate the code names in the CSV to English and store them
        in a JSON format in the CSV.
    3. Translations to other languages can be done through the
        Transifex plateform (danish and spanish is already done since
        they are the one using the codes).

task-3340441

closes odoo/enterprise#43994

Signed-off-by: Quentin De Paoli <qdp@odoo.com>