[FIX] mrp_workorder: shop floor sorting order of cards
Before this commit, the cards shown in the new shop floor display were
not sorted properly and did not get updated upon changing the workcenter
that was being displayed.
This commit will make sure that MO and WO cards are sorted by state and
start_date, and also refreshes the data in the display when we change
the workcenter that is currently shown.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44760

X-original-commit: d10fb8fa16ed2dcb32f18293f0292b6719ec14c0
Part-of: odoo/enterprise#45108