[IMP] mrp_workorder: make shop floor app optional
Before this commit, the shop floor app was always visible when the
mrp_workorder module was installed.
This commit adds a checkbox in the manufacturing settings to hide the
shop floor app from the home screen and the workorder form view.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44735

closes odoo/enterprise#44735

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>