[FIX] project_enterprise: update tour according to modules installed
Before this commit, the `project_enterprise_tour` has a step to
check allow billable during the project creation, this step can only
be done if `sale_project` module is installed but since
`project_enterprise` does not depend on that module the tour will
fail if only project_enterprise is installed.

This commit moves the project_enterprise_tour in project module
and rename the tour by `project_test_tour`. It also updates the
tour to be able to go to gantt and map views when `project_enterprise`
module is installed.

X-original-commit: 9a0bc2c96ded06e9339f17aa4e638d8a0e5b7f2a
Part-of: odoo/enterprise#44566