[FIX] web_studio: reportEditor: sync FieldDynamicPlaceholder position
Spawn the FieldDynamicPlaceholer in the report editor.
Scroll the container.

Before this commit, the FieldDynamicPlaceholder stayed at its birth position.
This is because the Popover Component synchronizes its position whenever the
anchor's document is scrolled.
In the reportEditor, the iframe's contents is not scrollable, but the iframe's container
is, so no scroll event is fired inside the iframe, and the Popover can't react properly.

This commit fixes this by manually firing a scroll event in the iframe whenever the container has scrolled

closes odoo/enterprise#44842

X-original-commit: a08493a2e9fecd8eba936292acbff0c1e7357c1a
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>