[FIX] sale_purchase_inter_company_rules: run tests
If only this module is installed, it will not be possible to run the
tests of the class `TestInterCompanySaleToPurchase`, because an
error is triggered when writing on `partner_invoice_id` in
`_generate_draft_sale_order`:
```
AssertionError: can't write on invisible field partner_invoice_id
```

The user has to be part of a specific group (`account.
group_delivery_invoice_address`) for this field to be displayed,
which is not the default configuration.

Moreover, the field is not required for the tests to be correctly
executed.

X-original-commit: 5459f99d53005eb720a10f1170e650b219e7d70f
Part-of: odoo/enterprise#45235