[FIX] account_asset: reevaluating asset
In 16.0 when creating an asset at a certain date, then create a lock date (with
a date after the asset), if the user want to modify the asset prior to the lock
date an error should be raised since the asset cannot be modified at an earlier
date than the lock date.

In this PR, we decided to add an error when the date of the modification of the
asset is prior the lock date, and also correcting the traceback to be more
robust.

Task: 3430603
Part-of: odoo/enterprise#44482