[IMP] *_extract: back to synchronous uploads
In Odoo 16, the upload of the documents to the extraction server was
moved to an asynchronous scheme through the usage of crons (see task
2888454).

While this looked like a good idea at first, it turned out that there
were some issues in practice with this solution:
- Even when triggering the cron instantly, we have no guarentee that it
  will run right away as the cron workers could be busy with other
  tasks.
- Low feedback to the user on the state of the upload as it was happening
  in the background.

The main reason for this change was that it improved the reactivity of
the interface as the upload itself occurred in the cron job.

Considering the situation, it has been decided that we'd go back to a
synchronous scheme as the cons mentionned above outweighted the pros.

Some improvements have been made to mitigate the latency due to the
re-introduction of synchronous uploads:
- On the OCR server side, the performances of the parsing routes have
  been improved to reduce its response time.
- An unnecessary call to get the IAP credits of the OCR service has been
  removed, this will be handled by the OCR server itself.

Task #3263472

Part-of: odoo/enterprise#44488
Co-authored-by: Louis Baudoux <lba@odoo.com>