[FIX] mrp_workorder: remove workcenter settings from demo
Since shop floor app. The mrp_workorder module is an auto install
of mrp and not trigger by the setting of work orders anymore.

Also the mrp_workorder contains now the shop floor application. The shop
floor app doesn't require the workorder and could be use for simple
use case with only production orders

mrp_workorder should be rename mrp_enterprise in a future work.
Currently since the workorders are not related to the module, we
remove it from the demo data

Part-of: odoo/enterprise#45699