[FIX] l10n_pe_edi: Handle 500 Server Error from SUNAT
Bugfix. When sending an invoice to SUNAT, if the CDR is already
registered on the server, SUNAT will respond with a 500 Server Error.
The response content indicates that the invoice was already registered
with other data.

At the moment, we bail out as soon as we see the 500 Server Error,
meaning that we don't check the response content and don't try to
retrieve the CDR from the server even though that would be appropriate.

With this commit, if we get a 500 Server Error, we still try to decode
the response content.

closes odoo/enterprise#44900

X-original-commit: 59cfadfabc246d782e4a359d365efb11b8af7ae7
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>