[FIX] sign: display company logo when signing doc
Steps to reproduce:
-------------------
1. Have at least two companies
2. Set a logo image for both companies
3. Use the second company (id = 2)
4. Go to Sign -> Upload a PDF to sign -> Send
5. Send the pdf by email
6. Go to Settings/technical/emails and search the last email you sent
7. Click on “Sign Document”
8. In the header of the website the logo is wrong, using the logo from company id=1

Issue:
------
When creating a sign document from any other company different than the main company, it doesn't take the logo of the company that is creating the document. For all documents it is taking the logo of the main company (the first company created in the database with id=1)

Cause:
------
The XML file is displaying the logo image from the path ‘/logo.png’ without specify the company_id. To fix this issue we need to pass the parameter ?company=<id> and we also have modify the controller to ensure that this parameter is passing correctly to the XML file.

OPW-3427766

closes odoo/enterprise#44500

Signed-off-by: Matheus Leal Viana (malv) <malv@odoo.com>