[FIX] web_studio: add aggregate option on float field
Since the rewrite of Studio in Owl, the aggregate option was unavailable
on float fields in list views. This commit brings back the possibility
to edit this option on the field of those types.

A test has been edited to verify the presence of the option when using
studio on a list view on monetary and float fields as well.

Finally, dead code related to the aggregation behavior has been removed
from list_editor_sidebar.js

task-3433586

closes odoo/enterprise#44626

X-original-commit: d003ad17d6ab6b4519af08e4ac8e55da76bfc038
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>