[FIX] web_studio: navigate in X2many list that has a record and context
Have a form view like:
```xml
 <form>
      <field name='product_ids'>
          <tree>
              <field name="m2o" context="{'context_key': 'value', 'parent': parent.id}" />
          </tree>
     </field>
 </form>
```

In studio, edit this view Form and navigate to edit the X2many List View. (provided that
the form's record has at least one record in its X2Many, which has its m2o field set).

Before this commit there was a crash as the new RelationalModel couldn't evaluate the "parent" key
in the context.

There was an asymmetry between the list and the form view in this flow, because we did not do
exactly the same thing to put "parent" in the context.
Namely, in the form view, we pushed that key in the config.context, which, happened
to be the same Object as params.context.

This commit is the occasion to harmonize all this and make this flow more robust.

closes odoo/enterprise#44534

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>