[FIX] knowledge: autofocus editor when opening an article
When someone opens an article in Knowledge, we would like to automatically
focus the editor, allowing people to start typing without the need to click.
The cursor will help users understand that the document is editable.

This commit will ensure that the editor of Knowledge will automatically
be focused when opening an article.

task-3416522

closes odoo/enterprise#46208

X-original-commit: 1dc478659a4ee3180e43d5a2ebdd9a7a28fa6eb4
Signed-off-by: Adrien Schoffeniels (adsc) <adsc@odoo.com>