[FIX] sign: fix deferred negative ids logic
https://github.com/odoo/enterprise/pull/43125 introduced a new relational model
and since sign uses it to show x2many records in the sign item popover when
editing a template. That PR also introduced a new approach to guarantee that
new sign items are already saved in the server before we allow opening the
popover. However, the logic for doing that was affecting items that already
had an ID set in the server because it was checking for the updated property in
the sign item, while it should be checking for a negative id only. This commit
fixes this by adding only the negative ids to the negativeIds object, which
contains promises for each id that are resolved when the item is saved in the
server.

task-3438789

closes odoo/enterprise#44551

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>