[FIX] l10n_pe_edi: allow preview invoice if national bank account is not present
When the user tries to preview an invoice by performing the
following steps, a traceback will be generated

Steps to reproduce:
1. Install the `l10n_pe_edi` module.
2. Switch to PE company
3. Go to invoice app.
4. Create a new invoice with `operation_type` in ['1001' to '1004']
5. Select a product in the `invoice lines` which have the value
  `Withhold Percentage` greater than 0, save it.
6. Click on `preview` button.

With the recent changes in commit [1] the condition is to check
`national_bank_account` has been removed because of that, the error is generated.

This commit fixes the above issue by allowing a preview of invoices when
the national bank account is not present.

commit [1] -
https://github.com/odoo/enterprise/commit/fd3ac630f50542668b456397bf08abc383e23011#diff-deaab3f5010fea8defc8af11dc186415ecc9079d86d307537b09222dda7ab751L176

sentry-4337938778

closes odoo/enterprise#45397

X-original-commit: a25f7663e6af4f82e7d8fbe1e5afa25c1a510fa8
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>