[FIX] mrp_workorder: don't open new QC when fail
In the Shop Floor app, before this commit, validating a Quality Check
will open the next QC. But in case of failed QC, there is often a popup
to alert the user or to ask confirmation or correction (like with
measure type's QC).
Those popups open at the same time the next QC is displayed which is a
bit messy.

So, to avoid this issue, this commit will open the next QC only when the
previous QC is good.

task-3432610

Part-of: odoo/enterprise#45367