[REF] web_enterprise,web_studio: remove legacy code
This PR is the counterpart of odoo/odoo#129314 where we remove the
support of legacy client actions.

This commit removes the override of the legacy ControlPanel, and
removes a legacy compatibility piece of code in the IconCreator,
which was no longer necessary, and which allows to remove the
legacy CustomFileInput component.

Part of task~3439226

closes odoo/enterprise#44465

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>