[FIX] mail_enterprise: stop WebClientEnterprise from leaking in tests
In order to add the push-notification capability to the PWA in
enterprise, the enterprise webclient needs to listen for changes in
permissions. The listener was not correctly cleared on destroy, which
caused tests that use the enterprise webclient to leak the entire owl
application and its corresponding DOM (this was mostly visible in studio
tests)

This commit fixes that by removing the event listener when the webclient
is destroyed.

closes odoo/enterprise#44653

X-original-commit: 5ba52f0785e0d208a44d368bccbde6db2c749508
Related: odoo/odoo#129813
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>