[FIX] stock_barcode: use the id of the package in the domain
Steps to reproduce the bug:
- Activate the packages and lot tracking
- Create a new storable product tracked by lot
- Add some stock for this product and give a lot and package number
- Print the lot (or package) barcode
- Open the barcode app and start a new inventory adjustment
- Try to create a new line by scanning the lot barcode

Problem:
A traceback is triggered: “psycopg2.ProgrammingError: can't adapt type
'dict'”

When scanning a lot, the function `_processBarcode` is called:

https://github.com/odoo/enterprise/blob/c0151bce3c60c69e7cadeb719a81c4a702b71c34/stock_barcode/static/src/models/barcode_model.js#L864

In this function, we check if a package has been scanned. If it is the
case, then the function `_processPackage` is triggered. Inside this
function, we call `_createNewLine` with the package ID, which allows
us to use it in the domain, and everything works perfectly:

https://github.com/odoo/enterprise/blob/c0151bce3c60c69e7cadeb719a81c4a702b71c34/stock_barcode/static/src/models/barcode_model.js#L891

https://github.com/odoo/enterprise/blob/c0151bce3c60c69e7cadeb719a81c4a702b71c34/stock_barcode/static/src/models/barcode_quant_model.js#L123

However, in this case, when scanning the lot, we have to check if
the package has been set, and if so, we are retrieving all the package
information into a dictionary instead of just adding his ID to the
barcodeData:

https://github.com/odoo/enterprise/blob/c0151bce3c60c69e7cadeb719a81c4a702b71c34/stock_barcode/static/src/models/barcode_model.js#L976

As a result, when the function `_createNewLine` is called, we try to
use all the package information (dict) into the domain, which generates
an error.

opw-3439193

closes odoo/enterprise#48120

X-original-commit: 5804b3bf56a4971b27aedafc1302805817cf5b9c
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>