[FIX] mrp_workorder: display by-product WO
In the Shop Floor app, display the by-product's lines only for the
corresponding Work Order if this by-product is marked as produced in the
WO's operation. If not linked to any operations, the by-product lines
will be displayed on every WO.

Also, avoid to display the move to register a by-product in the MO card
if there is a Quality Check to register this by-product.

task-3432610

Part-of: odoo/enterprise#45367