[FIX] sign: fix InitialsAllPagesDialog to show current role
Currently the InitialsAllPagesDialog is selecting the first role instead of the
current one. This is due to the fact that in the owl template, the selected
option is defined by the true equality operator in JS between a string and a
number. This commit fixes this by using the == operator instead of ===

task-3431069

closes odoo/enterprise#47513

X-original-commit: a66b347614c8a5255194cf240d292be6f3f6714f
Related: odoo/odoo#135774
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>