[IMP] {web,timesheet}_grid: add grid_component to use it in row title
Before this commit, the row/section title was rendered with components
in `fields` registry but we have to use `Record` component to be able
to use those components since `record` is needed.

This commit creates some grid components to be able to use it in
section/row title to avoid using the `fields` registry and Record
component in grid view.

task-3222154

Part-of: odoo/enterprise#44599