[FIX] industry_fsm, project_enterprise: patch gantt controller
Steps to reproduce:
-------------------
- go to Field Service app;
- go to Planning / By User tab;
- create a task.

Issue:
------
The default user will always be the user
and the default dates will be today's dates.

Cause:
------
The context will be used to create the task
on the server side to obtain the default values.

The three values that concern the problem
of this flow are: `rowId`, `start`, `stop`.

With the patch, we no longer have `rowId`
and therefore we no longer have the selected
frontend user.
In addition, `start` and `stop` are the times
of the current day's date.

Solution:
---------
Patch the controller used for the task gantt view.

opw-3420789

closes odoo/enterprise#45170

X-original-commit: 98834ee4e687308f8937fb0c06b90304eea01bfa
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>