[FIX] knowledge: fix error when reloading with no content helper
Purpose:
-------
When reloading the page when the no content helper is shown in the main
knowledge view, an error is shown before that the page is reloaded. This
happens because the quick create article is renamed in `beforeLeave` and
therefore tries to be created.
This commit prevents the automated renaming of the quick create article,
as the article form view is not handled in create mode (articles should
be created using the `article_create` method).

Task-3439418

closes odoo/enterprise#45285

X-original-commit: 99f04197b277a027c88a52fbdea87a4b8dd82f08
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>