[FIX] pos_blackbox_be: Do not allow product with negative price
Before it was possible to make a product with a negative price.
Thus, when using blackbox, it was doing a NR order.
The problem was that it was possible to mix positive and negative value of products, which is forbidden.

With this commit, we do not allow to add a product to the order with negative price.
The user is prompt with a message saying that he should use refund instead.

closes odoo/enterprise#44624

Related: odoo/odoo#129752
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>