[FIX] sale_subscription: fix sale cancel button is not visible
Before these commit when we create quotation and confirm it. we  will not be
able to see the cancel button on the form. 
Fix :
When we install the bridge module sale_subscription the cancel button gets hide
due to that we can't see the cancel button.
After doing some changes in sale_subscription we will be able to see the cancel
button on the form.

task-3433526

closes odoo/enterprise#45732

X-original-commit: b876b1a2806d0bc645ec4dc1521c767868ed0bbd
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>