[FIX] account_asset: reevaluating asset
In 16.0 when creating an asset at a certain date, then create a lock date (with a date after the asset), if the user want to modify the asset prior to the lock date an error should be raised since the asset cannot be modified at an earlier date than the lock date.

In this PR, we decided to add an error when the date of the modification of the asset is prior the lock date, and also correcting the traceback to be more robust.

closes odoo/enterprise#45150

Task: 3430603
Forward-port-of: #44482
X-original-commit: 42ac00da24b33ff2cd6c92a18120a970337121ef
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>