[FIX] web_studio: properly keep debug flag in tours
There is a commit in community that changes the way the debug mode works
for the web client: it now reads the value in the url instead of the
session (but only the web client). So, this means that tours that depend
on being in debug mode are no longer working, if they redirect to the
`/web` url without adding explicitely the `debug` flag.

See PR https://github.com/odoo/odoo/pull/129503

closes odoo/enterprise#44783

X-original-commit: d88ef17055f7396638ea08bbc182e264d03f9579
Related: odoo/odoo#130060
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Géry Debongnie <ged@odoo.com>