[FIX] web_studio: ReportEditor in multicompany
Before this commit, the company used to get all the report's data, including its QWeb
was the one set on the user, regardless of the company set by the companySwitcher component.

This was due to the fact that the "allowed_company_ids" key was not passed in the context.

After this commit, the current company is used to retrieve the report's data.

closes odoo/enterprise#44704

X-original-commit: d23fa177ef3a41b5a1835a09649eff62c73aadad
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>