[FIX] account_reports: prevent account taxes on grouped deferral moves
When we have default taxes set up for an account, we don't want the
deferral entries to apply these default taxes since that would impact
incorrectly impact our tax report.

We already fixed this issue for the automatic deferral method in
821e0e0af817d707cfe7729c463c78ef39a0ee35, but we forgot to adapt the
manual method. This commit applies the same fix to the manual method.

task-3441167

X-original-commit: 14b1e57b286d1c74d15c3e25986b7edcea2878a6
Part-of: odoo/enterprise#44702