[FIX] account_reports: link all grouped deferral moves
Currently when we generate the grouped deferral moves for one or more
moves, we link the generated deferral moves to the original move by
overwriting the previous value. When we consequently generate the
deferral moves for the next period, the linked deferral moves from the
previous period is not linked to the original move anymore.

This commit fixes the issue so that all grouped deferral moves are
linked to the original moves.

task-3441167

closes odoo/enterprise#44702

X-original-commit: adb3041cdd0504d3c168d831fda329088eeeddad
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>