[FIX] website_sale(_stock)_renting: buffer to allow immediate renting
Steps to reproduce:
- Go to /shop on a day where a pickup is allowed by the company;
- Select a rental product;
- Select the minimum available date (now, now + preparation time or
today, depending on the product setting);
- Add to cart;

Current behaviour:

The product is not added to the cart and a warning says that you cannot
rent in the past. Same check would prevent to pay for it.

Expected behaviour:

The product should be available to cart and payment should not be
blocked.

Note: we now add a buffer of 15 minutes between the moment the customer
adds a product and the check to fail, so that the customer can rent as
soon as possible.

OPW-3436743

closes odoo/enterprise#44727

X-original-commit: 962550cfceaf8975576b9e64d04f1e5afe27b02f
Signed-off-by: Vallaeys Valentin (vava) <vava@odoo.com>