[FIX] helpdesk: fix SLA short deadline outside workhours
Issue:
When we have an SLA policy with a duration shorter than the average
working hours for the calendar's team, and a ticket is created after
working hours, said SLA policy is viewed as failed upon ticket creation

Steps to reproduce:
- Install Helpdesk
- Create a Team > Activate SLA policies
- Create a Working Calendar whos working time is before 'now' (for
  ex: working hours in the morning, but ticket creation is in the
  afternoon)
- Set the calendar to the team
- Create an SLA policy of 2 hours for this team
- Create a new Ticket for said team, upon saving, it's SLA is
  already failed.

Cause:
In f2cb1712869a0c5cc0e8a0f5eac9814099cf9be5, we prevented
overshooting of the SLA deadline for days, but this shouldn't be
done when `status.sla_id.time < working_calendar.hours_per_day`,
because then we don't fall into `if time_days > 0:` which pushes the
deadline by 1 day in the future for that case. This means in our
use-case, the deadline is actually 'start of working day' + `status.
sla_id.time` of the same day, meaning that if the ticket was created
after the working day, it already exceeded the deadline.

Fix:
Add the condition to make sure the use-case referenced in the above
mentioned commit is used only in case of sla_times larger than a day.

Affected versions:
15.0 up to master

opw-3425366

closes odoo/enterprise#46660

X-original-commit: 295ef0fba639d61956f6e6d2a3fe48f000c8c83d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>