[FIX] timesheet_grid: 'All Timesheets' views editable with studio
Steps to reproduce
==================

- Go to Timesheets > All Timesheets
- Open studio
- Click on Views
- Disable the Pivot view and set the list view type as default
- Close studio, reload the page
- Nothing changes

Cause of the issue
==================

Studio edits the action view_mode. Since
0d18fe7c6ceb6fa0aee82d4e4f611f49eefc4993, the pivot view is always at
the start and the view_mode is ignored.

Solution
========

The goal of the previous commit is to only use the pivot view if there
is no type_view. Add back the pivot view to the view_mode. When we need
to remove it, simply filter it out. This means that we respect the
action view_mode and we allow editing it, but we can still remove the
pivot view when needed.

opw-3438086

closes odoo/enterprise#44645

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>