[FIX] account_followup: fixed plain text for followup email
Currently, the "Add a note" span is hidden from HTML view with a t-att-style.
Receivers of the follow up report who view their emails in plain text will still see the "Add a note" line.
The new change using t-if will properly hide the span for users that view the email in plain text.

opw-3413962

closes odoo/enterprise#45010

X-original-commit: ed4abcceee660de36a3ce04c9bc06a127f5b6795
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>