[FIX] mrp_workorder: shop floor enforce allowed employees
Before this commit, when only certain employees were allowed to work on
a workorder, it was still possible to modify the stock moves and
quality checks when the current employee was not allowed to do so, by
using the new shop floor display.
This commit disables the workorder cards in the new shop floor display
when the logged-in employee is not allowed to work on it.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44760

X-original-commit: a4f87022301b7fbe6cc6d2b01c51b04f24c7ffd6
Part-of: odoo/enterprise#45108