[FIX] knowledge: fix template command tour
Add a step to ensure that the Partner form view is fully loaded before clicking
on the Knowledge button in the chatter to search for articles. That button is
also present in the chatter of Knowledge article and if the Partner Form view
loading takes more than 100ms (a tour step), the tour could use that button in
the wrong Form view.

task-3442646

X-original-commit: 8440596c1182ddbb545a280bfc3d7d6a7d9ef931
Part-of: odoo/enterprise#45309