[FIX] spreadsheet_edition: allow model drilldown in global filters
The drilldown of fields in the global filters was only available to
users in debug mode when introduced in 93d78389. However, the feature
was kind of exposed to standard users due to the limitation of
`ModelFieldSelector`. It was decided that the feature could/should be
available for every user.

This commit also fixes the fact that we did not ensure that the target
field of the global filter was actually searchable.

opw 3435507

closes odoo/enterprise#44674

Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>