[FIX] mrp_workorder: WO validation
In the Shop Floor app, the validation of a MO or a WO usually take same
time to let the user cancels the validation if they want.
But while actually validating a MO, some wizard can happen, so to avoid
the user waiting the validation delay and then get a wizard, we check if
a wizard will popup by calling `pre_button_mark_done` before to add the
record into the validation stack.

That said, before this commit, we don't do that while validating a WO.
Since validate the last WO will also validate its MO, we call
`pre_button_mark_done` in this case.

task-3432610

Part-of: odoo/enterprise#45367