[FIX] account_reports: no deferrals for draft moves
When we have our deferred method set to "Manual & Grouped" and we have
a move that needs to be deferred, but is still in draft, we can show
the deferred report with the option "Show Draft entries".

If we do so, and we generate the deferral moves from that report, it
will also generate deferral moves for the draft moves included in the
report. We don't want that.

This change makes sure that deferrals are never created for draft moves.

task-3441779

closes odoo/enterprise#44710

X-original-commit: cb2a010e6aef8e73f69e1051a40a4007307d13fb
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>