[FIX] timesheet_grid: autocomplete dropdown with domain
Issue:
In the timesheet app, in the timer header, when we have a project
specified, the task dropdown shows all tasks, not only those
specified in the project we have selected.

Steps to reproduce:
- Install Timesheets
- Create a Project with a few tasks inside (make sure the project is
  billable)
- Go to Timesheets > Start the timer, select the project, notice
  that in the dropdown menu for the tasks we have all tasks.

Cause:
The component `timesheet_timer_header` doesn't make use of
Many2XAutocomplete component, so when doing the `name_search` and
fetching the domain for the field, `getDomain` return `undefined`.

Fix:
Add a `domain` props on each `Field` subcomponent of the
`TimesheetTimerHeader` component, which returns a function that
gives you the correct domain, which just fetches the domain from the
`fields` props which is a dictionary containing the our `fieldInfo`.

Affected versions:
saas-16.3 up to master

References:
opw-3418151
opw-3420312

closes odoo/enterprise#45564

X-original-commit: 7cf71b502aa4d02569da198a461ce9cedc4edf86
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>
Co-authored-by: thle-odoo <thle@odoo.com>